Styling, features, bugs...
First of all, prettier is finally enforced and now the code is relatively clean.

Secondly, users now have the ability to delete a note, although this is not enforced with permissions so any one who is logged in can delete another user's notes.
Fixed a problem from a previous commit that prefers upsert to the traditional find and update, this is weird because upsert was not working as expected. This is a workaround to #12.

Lastly, this commit fixes a bug that occurs when an user creates a note but decided to use the generated slug from the title. The composer now redirects to that note properly when it is created.

BONUS... Added better looking index page for Notes when there isn't any note selected. This is just easier on the eye.